Remove support for loading IMG, CCD, SUB files. Reduces DOL file size
Suggested by malvarenga123: https://github.com/xjsxjs197/WiiSXRX_2022/issues/139#issuecomment-2075589747

malvarenga123: To me that answer was really strange as the sub file is very small compared to the main image, so you shouldn't need to read it that frequently.
What I think is more plausible is that the general reading routine (for all formats) is not really optimized and the constant access to USB or SD card is causing those stutters.

xjsxjs197 already said that if using CCD/IMG/SUB, it causes many stutters not present on BIN/CUE, CHD, or other formats.
https://github.com/xjsxjs197/WiiSXRX_2022/issues/139#issuecomment-2071783259

Reduces DOL file size in 896 bytes.